# docs/api/openapi.v0.json
{
  "components": {
    "schemas": {
...
      "AccountStatus": {
        "enum": [
          "active",
          "inactive",
          "frozen",
          "erased"
        ],
        "type": "string"
      },
      "AccountType": {
        "enum": [
          "resident",
          "staff",
          "manager",
          "family"
        ],
        "type": "string"
      },
...
      "CreateAccountDto": {
        "properties": {
          "birthDate": {
            "description": "Required for a passport, ignored for a national ID.",
            "format": "date",
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "fullName": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string"
          },
          "idDocumentNumber": {
            "description": "14-digit national ID, or a passport number.",
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          },
          "idDocumentType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IdDocumentType"
              }
            ]
          },
          "nationality": {
            "description": "ISO 3166-1 alpha-2; required for a passport.",
            "type": "string"
          },
          "phone": {
            "description": "Local Egyptian format, or international with `+`; stored as E.164.",
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              }
            ]
          }
        },
        "required": [
          "type",
          "fullName",
          "idDocumentType",
          "idDocumentNumber",
          "phone",
          "email"
        ],
        "type": "object"
      },
...
      "DeletionConfirmationDto": {
        "properties": {
          "confirmation": {
            "description": "The word the holder types: `حذف` or `DELETE`.",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "type": "object"
      },
      "DeletionRequestResponse": {
        "properties": {
          "effectiveAt": {
            "description": "Undo is possible until then.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "cancelled",
              "completed"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "requestedAt",
          "effectiveAt"
        ],
        "type": "object"
      },
...
      "HouseholdSummaryView": {
        "properties": {
          "memberCount": {
            "type": "number"
          },
          "pendingInvites": {
            "type": "number"
          }
        },
        "required": [
          "memberCount",
          "pendingInvites"
        ],
        "type": "object"
      },
      "IdDocumentType": {
        "enum": [
          "national_id",
          "passport"
        ],
        "type": "string"
      },
...
      "Locale": {
        "enum": [
          "ar",
          "en"
        ],
        "type": "string"
      },
      "LocaleView": {
        "properties": {
          "preferredLocale": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Locale"
              }
            ]
          }
        },
        "required": [
          "preferredLocale"
        ],
        "type": "object"
      },
...
      "MeView": {
        "properties": {
          "email": {
            "nullable": true,
            "type": "string"
          },
          "fullName": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "idDocumentNumberMasked": {
            "example": "••••1234",
            "nullable": true,
            "type": "string"
          },
          "idDocumentType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IdDocumentType"
              }
            ],
            "nullable": true
          },
          "nationality": {
            "nullable": true,
            "type": "string"
          },
          "phone": {
            "nullable": true,
            "type": "string"
          },
          "preferredLocale": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Locale"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountStatus"
              }
            ]
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              }
            ]
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "fullName",
          "phone",
          "email",
          "preferredLocale",
          "idDocumentType",
          "idDocumentNumberMasked",
          "nationality"
        ],
        "type": "object"
      },
...
      "MyUnitView": {
        "properties": {
          "building": {
            "nullable": true,
            "type": "string"
          },
          "capacity": {
            "enum": [
              "owner",
              "tenant",
              "member"
            ],
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "floor": {
            "nullable": true,
            "type": "number"
          },
          "household": {
            "allOf": [
              {
                "$ref": "#/components/schemas/HouseholdSummaryView"
              }
            ],
            "description": "The primary only."
          },
          "isPrimary": {
            "type": "boolean"
          },
          "resides": {
            "description": "False only for an owner-landlord.",
            "type": "boolean"
          },
          "unitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "unitId",
          "code",
          "building",
          "floor",
          "capacity",
          "resides",
          "isPrimary"
        ],
        "type": "object"
      },
      "MyUnitViewList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MyUnitView"
            },
            "type": "array"
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "data",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "RevokedView": {
        "properties": {
          "revoked": {
            "type": "number"
          }
        },
        "required": [
          "revoked"
        ],
        "type": "object"
      },
...
      "SessionView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isCurrent": {
            "description": "The session of this request.",
            "type": "boolean"
          },
          "lastUsedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "userAgent": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "lastUsedAt",
          "userAgent",
          "isCurrent"
        ],
        "type": "object"
      },
      "SessionViewList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SessionView"
            },
            "type": "array"
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "data",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "UpdateAccountStatusDto": {
        "properties": {
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "UpdateLocaleDto": {
        "properties": {
          "locale": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Locale"
              }
            ]
          }
        },
        "required": [
          "locale"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Jiwar compound-management platform. API v0 is a draft: endpoints are reshaped screen by screen with the design, and every operation carries `x-stability: draft` (ADR 0025).",
    "title": "Jiwar API",
    "version": "0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "operationId": "AccountsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ],
        "x-stability": "draft"
      },
      "post": {
        "operationId": "AccountsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/accounts/{id}/status": {
      "patch": {
        "operationId": "AccountsController_updateStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountStatusDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ],
        "x-stability": "draft"
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
...
    "/api/v1/me/deletion-request": {
      "get": {
        "operationId": "MeController_deletionRequest",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionRequestResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      },
      "post": {
        "operationId": "MeController_requestDeletion",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletionConfirmationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionRequestResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/deletion-request/cancel": {
      "post": {
        "operationId": "MeController_cancelDeletion",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/locale": {
      "patch": {
        "operationId": "MeController_locale",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocaleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocaleView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/sessions": {
      "get": {
        "operationId": "MeController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionViewList"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/sessions/revoke-all": {
      "post": {
        "operationId": "MeController_revokeAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/sessions/{id}/revoke": {
      "post": {
        "operationId": "MeController_revoke",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
    "/api/v1/me/units": {
      "get": {
        "operationId": "MeUnitsController_units",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyUnitViewList"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ],
        "x-stability": "draft"
      }
    },
...
  },
  "servers": [],
...
}